KD-1853, Migrate Atomicupdater to new Koha version
When deploying production databases to test environments, it is imperative to
keep track of which database changes have been applied.

When one is running, in production, features that require DB changes, and want
to add more features which need more DB changes, keeping track of which updates
have been applied is challenging.
After several cycles of upgrade-deploy, with multiple atomicupdate-scripts
piling to the atomicupdates/-directory, it is impossible to easily distinguish
which updates have been applied and which are not.

Rerunning the same update scripts again and
again causes lots of noise in the upgrade log and hides real issues from being
detected.
Also repeatedly running upgrade scripts might cause bad side effects which are
potentially hard to repair.

This feature adds a script atomicupdate.pl which runs all atomicupdate/* scripts
and .sqls and logs the event to the koha.atomicupdates-table.
On subsequent runs of atomicupdate.pl, the already deployed upgrades are
skipped, greatly reducing the unnecessary clutter and risk of upgrading.

Also the order in which the atomicupdates need to be applied is important.
This _updateorder is automatically generatable from the original Git repository.

You can also remove existing log entries to allow rerunning the same upgrade
script again, list all applied upgrades and show pending updates.

Unit tests included.

TEST PLAN:

1. Run "perl installer/data/mysql/atomicupdate.pl --help"
2. Read the code.
3. git bz apply some bugs which introduce atomicupdates, or cook your own
4. Deploy the atomicupdates using the atomicupdate.pl-script.
5. Repeat steps 3-4 to your heart's content.
6+. Understand the code and come up with improvement ideas!